requirements.txt
python-dotenv==0.20.0
pyTelegramBotAPI==4.6.0